requirements.txt
Authlib==1.3.2
h2==4.1.0
litellm==1.50.0
pillow==11.0.0
psycopg2-binary==2.9.10
pydantic-settings==2.6.0
reflex-clerk==1.0.3
supabase==2.9.1
